Move contexts to unversioned package
